Revert "refactor: Moving tests around to match Preact's style"
This reverts commit 1577c8fe8591d21059a27646146987f9d72de1a7.
