+++ src/__test__/index.spec.tsx
@@ -1,8 +1,8 @@
 import { route } from 'preact-router'
 import { h, render } from 'preact'
 import { unmountComponentAtNode } from 'preact/compat'
-import App from '../src/App'
-import { request } from '../src/services'
+import App from '../App'
+import { request } from '../services'
 
 jest.mock('preact-router')
