+++ src/components/__test__/ArticlePreview.spec.tsx
@@ -1,11 +1,10 @@
-import { h } from 'preact'
+import { generateArticles } from '../../utils/test-utils'
+import ArticlePreview from '../ArticlePreview'
 import { shallow } from 'enzyme'
+import { h } from 'preact'
+import { deleteFavoriteArticle, postFavoriteArticle } from '../../services'
 
-import { generateArticles } from '../utils/test-utils'
-import ArticlePreview from '../../src/components/ArticlePreview'
-import { deleteFavoriteArticle, postFavoriteArticle } from '../../src/services'
-
-jest.mock('../../src/services')
+jest.mock('../../services')
 
 const setArticle = jest.fn()
