+++ src/components/__test__/Footer.spec.tsx
@@ -1,7 +1,6 @@
-import { h } from 'preact'
 import render from 'preact-render-to-string'
-
-import Footer from '../../src/components/Footer'
+import Footer from '../Footer'
+import { h } from 'preact'
 
 describe('# Footer', () => {
   it('should be display normally', function () {
