+++ src/components/__test__/NavBar.spec.tsx
@@ -1,11 +1,10 @@
+import { shallow } from 'enzyme'
+import NavBar from '../NavBar'
 import { h } from 'preact'
 import render from 'preact-render-to-string'
-import { shallow } from 'enzyme'
+import { useRootState } from '../../store'
 
-import NavBar from '../../src/components/NavBar'
-import { useRootState } from '../../src/store'
-
-jest.mock('../../src/store')
+jest.mock('../../store')
 
 const useRootStateMock = useRootState as jest.Mock
