+++ src/pages/__test__/Login.spec.tsx
@@ -1,17 +1,16 @@
 import { h } from 'preact'
+import Login from '../Login'
+import { postLogin } from '../../services'
 import { route } from 'preact-router'
 import { mount, shallow } from 'enzyme'
+import { setInputValue } from '../../utils/test-utils'
+import { useRootState } from '../../store'
+import { login } from '../../store/actions'
 
-import Login from '../../src/pages/Login'
-import { postLogin } from '../../src/services'
-import { setInputValue } from '../utils/test-utils'
-import { useRootState } from '../../src/store'
-import { login } from '../../src/store/actions'
-
-jest.mock('../../src/services')
+jest.mock('../../services')
 jest.mock('preact-router')
-jest.mock('../../src/store/actions')
-jest.mock('../../src/store')
+jest.mock('../../store/actions')
+jest.mock('../../store')
 
 const useRootStateMock = useRootState as jest.Mock
